feat: update repo to implement maxGap

+++ SBA_M/src/main/java/com/example/SBA_M/dto/request/UniversityMajorAdmissionRequest.java
@@ -12,5 +12,5 @@ import lombok.experimental.FieldDefaults;
 public class UniversityMajorAdmissionRequest {
     private Double score;
     private Long subjectCombinationId;
-    private Integer maxGap;
+    private Double maxGap;
 }
