+++ SBA_M/src/main/java/com/example/SBA_M/service/UniversityMajorService.java
@@ -31,7 +31,7 @@ public interface UniversityMajorService {
     List<UniversitySubjectCombinationSearchResponse> searchBySubjectCombination (
             Long subjectCombinationId, @Nullable String universityName) throws IOException;
     List<UniversityMajorResponse> getAllUniversityMajors();
-    List<UniversityMajorAdmissionResponse> findEligibleMajors(Double score, Long subjectCombinationId, Integer maxGap);
+    List<UniversityMajorAdmissionResponse> findEligibleMajors(Double score, Long subjectCombinationId, Double maxGap);
 
     List<UniversityMajorSearchResponse> searchByMajor (
             Long majorId, @Nullable String universityName) throws IOException;
